+++ app/services/users/index.js
@@ -1,0 +1,3 @@
+const { walletAddressExists } = require('./walletAddressExists')
+
+module.exports = { walletAddressExists }