+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## v1.0.0 ()
+
+*   Wallet address uniqueness
+*   Payload expiration checking
+
 ## v0.2.0 (Dec 21, 2022)
 
 *   Logout implementation